site/requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
blinker==1.4
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
decorator==4.4.2
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
ffmpeg==1.4
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
kaitaistruct==0.10
kombu==5.4.0
moviepy==1.0.3
mysqlclient==2.2.4
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
playwright==1.46.0
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
pyasn1==0.6.0
pycparser==2.22
pyee==11.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.7.0
redis==5.0.8
requests==2.32.3
scipy==1.14.1
selenium==4.23.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.20.1
zstandard==0.23.0
